Criação da classe controller, dos endpoints e alterações na classe service para se adequar às regras de negócio.

+++ src/main/java/com/brunoqueiroz/testedev/services/PlaylistService.java
@@ -1,8 +1,9 @@
 package com.brunoqueiroz.testedev.services;
 
+import com.brunoqueiroz.testedev.dtos.PlaylistDTO;
 import com.brunoqueiroz.testedev.repository.PlaylistRepository;
 import com.brunoqueiroz.testedev.model.Playlist;
-import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.beans.BeanUtils;
 
 import java.util.List;
 
@@ -18,12 +19,19 @@ public class PlaylistService {
     public Playlist findById(Long id) { return playlistRepository.findById(id).get(); }
 
 
-    public Playlist save(Playlist playlist) { return playlistRepository.save(playlist);}
+    public Playlist save(PlaylistDTO playlistDto) {
+        if(existsByName(playlistDto.nome())) throw new IllegalArgumentException("400 Bad Request");
+
+        Playlist playlist = new Playlist();
+        BeanUtils.copyProperties(playlistDto, playlist);
+        return playlistRepository.save(playlist);
+    }
 
 
-    public String delete(Long id){
+    public void delete(String name){
+        if(!existsByName(name)) throw new NullPointerException("404 Not Found");
+        Long id = playlistRepository.findByName(name).getId();
         playlistRepository.deleteById(id);
-        return "playlist deleted";
     }
 
     public Boolean existsByName(String nome){
@@ -31,6 +39,7 @@ public class PlaylistService {
     }
 
     public Playlist findByName(String nome){
+        if(!existsByName(nome)) throw new NullPointerException("404 Not Found");
         return playlistRepository.findByName(nome);
     }
 }
